Sticky report header rows
For reports that can have many rows, particularity treatments and daily stats, I have added css to sticky the header row in place at the top of the page as the user scrolls down the page.  It is helpful to keep header names in view.

I have added the grey background colour that was used in the treatments header row to both daily stats and weekly distribution header rows, as for all three it hides the table rows that pass under the header row

![treatments-sticky-header-row](https://user-images.githubusercontent.com/217953/104672803-60175a80-56d8-11eb-9e4e-f3f72595c161.png)

![weekly-dist-sticky-header-row](https://user-images.githubusercontent.com/217953/104672827-6dcce000-56d8-11eb-8a87-5f5eda0da3b4.png)

![daily-sticky-header-row](https://user-images.githubusercontent.com/217953/104672847-71f8fd80-56d8-11eb-9765-1dffc5edca08.png)

I have limited the sticky header rows to on-screen viewing.